+++ apis/productos/src/main/java/com/perfulandia/productos/repository/MarcaRepository.java
@@ -3,5 +3,5 @@ package com.perfulandia.productos.repository;
 import com.perfulandia.productos.models.Marca;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MarcaRepository extends JpaRepository<Marca, Long> {
+public interface MarcaRepository extends JpaRepository<Marca, Integer> {
 }
